Consolidate and clean up queries
